all: Make the block number in EthereumBlockPointer a BlockNumber
We can't handle u64 block numbers - any interaction with the store with a
block number that does not fit into an i32 will blow up. Rather than do the
conversion of u64 to BlockNumber adhoc and in many places, force the code
to do those conversions when we take in a block number from an external
source.

The code could be tightened up further by using a newtype for BlockNumber,
but that is left for another change.

+++ graphql/src/execution/execution.rs
@@ -365,7 +365,7 @@ pub async fn execute_root_selection_set<R: Resolver>(
             // JSONB and metadata queries use `BLOCK_NUMBER_MAX`. Ignore this case for two reasons:
             // - Metadata queries are not cacheable.
             // - Caching `BLOCK_NUMBER_MAX` would make this cache think all other blocks are old.
-            if block_ptr.number != BLOCK_NUMBER_MAX as u64 {
+            if block_ptr.number != BLOCK_NUMBER_MAX {
                 // Calculate the hash outside of the lock
                 let cache_key = cache_key(&ctx, &selection_set, &block_ptr);
                 let shard = (cache_key[0] as usize) % QUERY_BLOCK_CACHE.len();
